Add basic_dynamic_body.hpp
fix #606
