+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 80:
 
 * Javadoc tidying
+* Add basic_dynamic_body.hpp
 
 --------------------------------------------------------------------------------
