+++ test/http/CMakeLists.txt
@@ -13,6 +13,7 @@ add_executable (http-tests
     message_fuzz.hpp
     test_parser.hpp
     ../../extras/beast/unit_test/main.cpp
+    basic_dynamic_body.cpp
     basic_parser.cpp
     buffer_body.cpp
     doc_examples.cpp
