+++ test/http/Jamfile
@@ -7,6 +7,7 @@
 
 unit-test http-tests :
     ../../extras/beast/unit_test/main.cpp
+    basic_dynamic_body.cpp
     basic_parser.cpp
     buffer_body.cpp
     doc_examples.cpp
